Added utilities to handle common functions
