+++ src/components/PetDetail.jsx
@@ -1,5 +1,6 @@
 import PetAttributeRows from "./PetAttributeRows";
 import PetThumbnail from "./PetThumbnail";
+import { ReturnFormattedDate } from "../utilities/DateFunctions";
 import React from "react";
 
 const PetDetail = (props) => {
@@ -37,7 +38,7 @@ const PetDetail = (props) => {
             attributes={attributes}
             animalId={animalId}
             gender={gender}
-            dateAdoptedRedeemed={dateAdoptedRedeemed}
+            dateAdoptedRedeemed={ReturnFormattedDate(dateAdoptedRedeemed)}
             petType={petType}
           />
         </div>
