+++ src/utilities/ConvertToFriendlyNames.js
@@ -1,0 +1,17 @@
+export const ConvertToFriendlyNames = (value) => {
+  var newValue = "";
+  switch (value) {
+    case "Primary Breed":
+      newValue = "Breed";
+      break;
+    case "Altered":
+      newValue = "Spayed / Neutered";
+      break;
+    case "Primary Color":
+      newValue = "Color";
+      break;
+    default:
+      newValue = value;
+  }
+  return newValue;
+};
